+++ build.gradle
@@ -28,7 +28,8 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+//    mainClass.set("seedu.duke.Duke")
+    mainClass.set("rocket.Rocket")
 }
 
 shadowJar {
