+++ src/main/java/rocket/Command.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 public class Command {
