+++ src/main/java/rocket/Deadline.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
