+++ src/main/java/rocket/DeadlineCommand.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 import java.time.LocalDateTime;
