+++ src/main/java/rocket/DeleteCommand.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 public class DeleteCommand extends Command{
     public DeleteCommand() {
