+++ src/main/java/rocket/Event.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
