+++ src/main/java/rocket/RocketException.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 public class RocketException extends Exception {
     public RocketException(String message) {
