+++ src/main/java/rocket/Task.java
@@ -1,4 +1,4 @@
-package Rocket;
+package rocket;
 
 public class Task {
     //Solution below adapted from: https://nus-cs2103-ay2324s1.github.io/website/schedule/week2/project.html#:~:text=Partial-,solution,-Duke%20Level%2D4
