+++ src/main/java/rocket/TaskList.java
@@ -1,9 +1,4 @@
-package Rocket;
-
-import Rocket.Deadline;
-import Rocket.Event;
-import Rocket.RocketIllegalArgumentException;
-import Rocket.Task;
+package rocket;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
