+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-    Hello! I'm Rocket
+    Hello! I'm rocket
     What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
